% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{build_family_catalog}
\alias{build_family_catalog}
\title{Build catalog of term families}
\usage{
build_family_catalog(ref_level, non_ref_level, timepoint_var)
}
\arguments{
\item{ref_level}{Reference diagnosis level}

\item{non_ref_level}{Non-reference diagnosis level}

\item{timepoint_var}{Timepoint variable name}
}
\value{
Tibble with family patterns
}
\description{
Build catalog of term families
}
\keyword{internal}
