% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{create_empty_plot}
\alias{create_empty_plot}
\title{Create empty plot with message}
\usage{
create_empty_plot(message)
}
\arguments{
\item{message}{Message to display}
}
\value{
ggplot object
}
\description{
Create empty plot with message
}
\keyword{internal}
