% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print.longitudinal_diagnosis}
\alias{print.longitudinal_diagnosis}
\title{Print method for longitudinal_diagnosis}
\usage{
\method{print}{longitudinal_diagnosis}(x, ...)
}
\description{
Print method for longitudinal_diagnosis
}
\keyword{internal}
